Authorization Implemented

+++ src/api/notes/handler.js
@@ -15,22 +15,23 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
-
-    const noteId = await this._service.addNote({ title, body, tags });
+    const { id: owner } = request.auth.credentials;    
+    const noteId = await this._service.addNote({ title, body, tags, owner });
 
     const response = h.response({
       status: 'success',
       message: 'Catatan berhasil ditambahkan',
       data: {
-        noteId,
+        noteId, 
       },
     });
     response.code(201);
     return response;    
   }
  
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: owner } = request.auth.credentials;
+    const notes = await this._service.getNotes(owner);
     return {
       status: 'success',
       data: {
@@ -41,6 +42,9 @@ class NotesHandler {
  
   async getNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
     return {
       status: 'success',
@@ -53,7 +57,9 @@ class NotesHandler {
   async putNoteByIdHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
 
+    await this._service.verifyNoteOwner(id, credentialId);
     await this._service.editNoteById(id, request.payload);
 
     return {
@@ -64,6 +70,9 @@ class NotesHandler {
  
   async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
     await this._service.deleteNoteById(id);
 
     return {
